Added deprecated warning, created new base class, updated Causal and T5
