+++ llmfoundry/models/hf/__init__.py
@@ -9,8 +9,10 @@ from llmfoundry.models.hf.hf_fsdp import (
 )
 from llmfoundry.models.hf.hf_t5 import ComposerHFT5
 from llmfoundry.models.hf.model_wrapper import HuggingFaceModelWithFSDP
+from llmfoundry.models.hf.hf_base import BaseHuggingFaceModel
 
 __all__ = [
+    'BaseHuggingFaceModel',
     'ComposerHFCausalLM',
     'ComposerHFT5',
     'prepare_hf_causal_lm_model_for_fsdp',
